refactor neo4j query functions into a separate file
